口コミ機能追加

+++ src/database/migrations/2024_07_16_111654_add_roles_to_users_table.php
@@ -9,16 +9,24 @@ class AddRolesToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->boolean('is_admin')->default(false);
-            $table->boolean('is_store_representative')->default(false);
+            if (!Schema::hasColumn('users', 'is_admin')) {
+                $table->boolean('is_admin')->default(false);
+            }
+            if (!Schema::hasColumn('users', 'is_store_representative')) {
+                $table->boolean('is_store_representative')->default(false);
+            }
         });
     }
 
     public function down()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('is_admin');
-            $table->dropColumn('is_store_representative');
+            if (Schema::hasColumn('users', 'is_admin')) {
+                $table->dropColumn('is_admin');
+            }
+            if (Schema::hasColumn('users', 'is_store_representative')) {
+                $table->dropColumn('is_store_representative');
+            }
         });
     }
 }
