+++ src/database/migrations/2024_07_19_004313_add_shop_id_to_users_table.php
@@ -9,16 +9,20 @@ class AddShopIdToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedBigInteger('shop_id')->nullable()->after('id');
-            $table->foreign('shop_id')->references('id')->on('shops')->onDelete('cascade');
+            if (!Schema::hasColumn('users', 'shop_id')) {
+                $table->unsignedBigInteger('shop_id')->nullable()->after('id');
+                $table->foreign('shop_id')->references('id')->on('shops')->onDelete('cascade');
+            }
         });
     }
 
     public function down()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropForeign(['shop_id']);
-            $table->dropColumn('shop_id');
+            if (Schema::hasColumn('users', 'shop_id')) {
+                $table->dropForeign(['shop_id']);
+                $table->dropColumn('shop_id');
+            }
         });
     }
 }
